requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.5
cffi==1.11.5
cryptography==2.4.2
Django==2.1.5
django-paypal==0.5.0
Fabric3==1.14.post1
idna==2.8
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pytz==2018.9
selenium==3.12.0
six==1.12.0
virtualenv==16.2.0
